Add attention point with pagination
There is a attention point when you are using the gem and paginating the result (exclusively with postgres). The problem 
is not caused by the gem, but by postgres (when pagination is ordered with two or more records with the same order value). 

The issue is described on this StackOverflow 
[issue](http://stackoverflow.com/questions/13580826/postgresql-repeating-rows-from-limit-offset). 

I just documented the fact to help future users to understand.